refactor : 타입 파일명 professor.ts로 통일
