+++ src/entities/professor/model/Professor.ts
